Refactor user login system
Previously, we were using a `Maybe UserHome.Model` to determine the
logged-in/logged-out state of the app.

In this patch, we create a new type for the model to represent the user
status: `LoggedOut` or `LoggedIn User`. This provides a more
semantically-relevant indicator of the user status than `Maybe` did.

It also allows for some future simplification of the user code, since we
don't need to be storing that model as a `Maybe`.