Refactor login task out of login screen and into App
This completes the refactoring of the side-effecting tasks into the top
app component. Previously the actual API calls were being made in the
display components, which was something I wanted to avoid. Now, the
update functions have been changed to pass commands around to indicate
login intention and the main app actually performs the task and updates
itself.

This was significantly cleaned up some of the previous code that was
intercpeting child messages.