Refactor User.avatarUrl: Maybe String -> String (defaulted)
Previously we stored a `Maybe String` for the user's avatar, so that if
no avatar were set, we could record that. However, in practice, anywhere
that the avatar was unavailable, we would be pulling in some default,
such as a gravatar.

Now, we set a sane default on decode of the user so that there will
always be a valid URL available.