Refactor logged-out-screen into a flat model
This piece of the app will be a kind of sub-app composing the login
screen and the signup screen, as well as the lost-password or forgotten
login details tools.

This update reflects a transition towards more idiomatic Elm code,
removing a level of message wrapping/mapping.

A further PR should split the user home into similar sub-apps: the
messages view, search, shares, and matches.